Move the `bad_bias` check to main()
The original `bad_bias` check logic was done in track_super_photon().
The `bad_bias` flag is not check until the very end right before
outputing the final spectrum.  Therefore, compute is wasted while no
data is output.  If the decision is to abort when `bad_bias` is set
true, it is better to abort earlier so, e.g., workflow managers can
adjust the parameters and resume the job.